formatted file info
