+++ src/hooks/UnitHooks.js
@@ -1,0 +1,36 @@
+const formatTime = (mediaInfo) => {
+  const timestamp = mediaInfo;
+  const date = new Date(timestamp);
+  const milliseconds = date.getTime();
+
+  const elapsedMilliseconds = Date.now() - milliseconds;
+  const elapsedSeconds = Math.floor(elapsedMilliseconds / 1000);
+  const elapsedMinutes = Math.floor(elapsedSeconds / 60);
+  const elapsedHours = Math.floor(elapsedMinutes / 60);
+
+  if (elapsedSeconds < 60) {
+    return elapsedSeconds + 's';
+  }
+  if (elapsedMinutes < 60) {
+    return elapsedMinutes + 'm';
+  }
+  if (elapsedHours < 24) {
+    return elapsedHours + 'h';
+  }
+  const options = {day: 'numeric', month: 'short'};
+  return date.toLocaleDateString('en-US', options);
+};
+
+const formatSize = (sizeInBytes) => {
+  if (sizeInBytes < 1024) {
+    return sizeInBytes + ' B';
+  }
+  const sizeInKiloBytes = sizeInBytes / 1024;
+  if (sizeInKiloBytes < 1024) {
+    return sizeInKiloBytes.toFixed(2) + ' KB';
+  }
+  const sizeInMegaBytes = sizeInKiloBytes / 1024;
+  return sizeInMegaBytes.toFixed(2) + ' MB';
+};
+
+export {formatTime, formatSize};
